+++ src/burn/engine/plan.cpp
@@ -1814,7 +1814,10 @@ extern "C" HRESULT PlanSetResumeCommand(
 {
     HRESULT hr = S_OK;
 
-    // build the resume command-line.
+    // Build the resume command-line.
+    hr = StrAllocFormatted(&pRegistration->sczResumeCommandLine, L"\"%ls\"", pRegistration->sczCacheExecutablePath);
+    ExitOnFailure(hr, "Failed to copy executable path to resume command-line.");    
+
     hr = CoreRecreateCommandLine(&pRegistration->sczResumeCommandLine, action, pCommand->display, pCommand->restart, pCommand->relationType, pCommand->fPassthrough, pRegistration->sczActiveParent, pRegistration->sczAncestors, pLog->sczPath, pCommand->wzCommandLine);
     ExitOnFailure(hr, "Failed to recreate resume command-line.");