+++ src/burn/engine/pseudobundle.h
@@ -40,7 +40,7 @@ HRESULT PseudoBundleInitialize(
 HRESULT PseudoBundleInitializePassthrough(
     __in BURN_PACKAGE* pPassthroughPackage,
     __in BOOTSTRAPPER_COMMAND* pCommand,
-    __in_z_opt LPCWSTR wzApppendLogPath,
+    __in_z_opt LPCWSTR wzAppendLogPath,
     __in_z_opt LPWSTR wzActiveParent,
     __in_z_opt LPWSTR wzAncestors,
     __in BURN_PACKAGE* pPackage